requirements.txt
langchain==0.0.184
transformers==4.28.1
faiss-cpu==1.7.4
pytesseract==0.3.10
opencv-python==4.7.0.72
fastapi==0.95.2
streamlit==1.22.0
torch==1.13.1
torchvision==0.14.1
sentence-transformers==2.2.2
python-multipart==0.0.6
uvicorn==0.22.0
Pillow==9.5.0
PyPDF2==3.0.1
numpy==1.26.4
scikit-learn==1.5.2
langdetect==1.0.9
pydantic==1.10.18
huggingface-hub==0.25.2
tqdm==4.66.5
nltk==3.9.1
regex==2024.9.11
requests==2.32.3
joblib==1.4.2
packaging==24.1
filelock==3.16.1
urllib3==2.2.3
certifi==2024.8.30
idna==3.10
charset-normalizer==3.4.0
typing_extensions==4.12.2
asyncio==3.4.3